Centralizing the map data structures in one place
